refactor: styles file
