+++ features/rewards/features/top-card/styles.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import { WalletCardStyle } from 'shared/wallet/card/styles';
+
+export const WalletStyle = styled(WalletCardStyle)`
+  background: linear-gradient(
+    52.01deg,
+    #37394a 0%,
+    #363749 0.01%,
+    #40504f 100%
+  );
+  padding: 0 0 24px 0;
+`;
