+++ features/rewards/features/top-card/wallet.tsx
@@ -1,5 +1,4 @@
 import { FC } from 'react';
-import styled from 'styled-components';
 import { ThemeProvider, themeDark } from '@lidofinance/lido-ui';
 
 import { InputDescription } from 'features/rewards/components/inputDescription';
@@ -7,20 +6,10 @@ import { AddressInput } from 'features/rewards/components/addressInput';
 import { InputWrapper } from 'features/rewards/components/inputWrapper';
 import { useRewardsHistory } from 'features/rewards/hooks';
 
-import { WalletCardStyle } from 'shared/wallet/card/styles';
+import { WalletStyle } from './styles';
 
 const INPUT_DESC_TEXT =
   'Current balance may differ from last balance in the table due to rounding.';
-
-const WalletStyle = styled(WalletCardStyle)`
-  background: linear-gradient(
-    52.01deg,
-    #37394a 0%,
-    #363749 0.01%,
-    #40504f 100%
-  );
-  padding: 0 0 24px 0;
-`;
 
 export const Wallet: FC = () => {
   const {
