+++ shared/wallet/unsupported-chain-button/styles.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+import { Button } from '@lidofinance/lido-ui';
+
+export const ButtonStyle = styled((props) => <Button {...props} />)`
+  background: var(--lido-color-backgroundSecondary);
+  color: var(--lido-color-textSecondary);
+`;
